Add updateclient method
