+++ pkg/auth0/types.go
@@ -1,5 +1,11 @@
 package auth0
 
+type Client struct {
+	Callbacks         []string `json:"callbacks"`
+	AllowedLogoutURLs []string `json:"allowed_logout_urls"`
+	WebOrigins        []string `json:"web_origins"`
+}
+
 type ResouceServer struct {
 	Name                                      string `json:"name"`
 	Identifier                                string `json:"identifier"`
